"Added functionality to get input value and update timer display on button click"
This commit adds code to get the value of the input field where the user specifies the time for the timer. It also adds an event listener to the "Iniciar" button, so that when it is clicked, the timer display is updated with the value from the input field.

Additionally, the commit changes the "Iniciar" and "Pausar" buttons from simple HTML buttons to JavaScript objects, so that they can have event listeners added to them.